removed duplicate method from GTMath. Refactored related methods into GTMatrixUtils. Added javadoc for unclear methods.
