+++ src/main/java/com/gregtechceu/gtceu/api/recipe/chance/boost/ChanceBoostFunction.java
@@ -2,7 +2,8 @@ package com.gregtechceu.gtceu.api.recipe.chance.boost;
 
 import com.gregtechceu.gtceu.api.GTValues;
 import com.gregtechceu.gtceu.api.recipe.content.Content;
-import com.gregtechceu.gtceu.utils.GTMath;
+
+import net.minecraft.util.Mth;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -19,7 +20,7 @@ public interface ChanceBoostFunction {
         int tierDiff = chanceTier - recipeTier;
         if (tierDiff <= 0) return entry.chance; // equal or invalid tiers do not boost at all
         if (recipeTier == GTValues.ULV) tierDiff--; // LV does not boost over ULV
-        return GTMath.clamp(entry.chance + (entry.tierChanceBoost * tierDiff), 0, entry.maxChance);
+        return Mth.clamp(entry.chance + (entry.tierChanceBoost * tierDiff), 0, entry.maxChance);
     };
 
     /**
